Broke down the fp/ directory into subdirectories of modules ordered by type
