+++ src/colors/hue.ts
@@ -2,7 +2,7 @@
 // @ts-nocheck
 
 import { getChannel } from '../getters_and_setters/get.ts';
-import { sortedArr } from '../fp/array.ts';
+import { sortedArr } from '../fp/array/sortedArr.ts';
 import type { Color, HueColorSpaces, factor } from '../paramTypes';
 
 //  Globals
