+++ src/filterBy/filterByLightness.ts
@@ -1,4 +1,4 @@
-import { filteredArr } from '../fp/array.ts';
+import { filteredArr } from '../fp/array/filteredArr.ts';
 import { getChannel } from '../getters_and_setters/get.ts';
 import { Color, factor } from '../paramTypes.ts';
 /**
