+++ src/fp/array/colorObjArr.ts
@@ -1,0 +1,20 @@
+// @ts-nocheck
+
+import { colorObj } from '../object/colorObj';
+import { factor, Color } from '../../paramTypes';
+
+/*
+ * @function
+ * @private
+ * Creates a custom object with a factor to pass to the predicate function.
+ * @param factor The quality being queried.
+ * @param cb The callback function for computing the factor's start.
+ * @param colors The array of colors to iterate over.
+ * @returns An array of objects.
+ */
+export const colorObjArr =
+  (factor: factor, callback) =>
+  (colors: Color[]): Array<{ factor: factor; name: Color }> => {
+    const cb = colorObj(factor, callback);
+    return colors.map((color) => cb(color));
+  };
