+++ src/fp/array/index.ts
@@ -1,0 +1,7 @@
+// This module has array methods
+// @ts-nocheck
+export { min, max } from './min_max';
+export { colorObjArr } from './colorObjArr';
+export { customSort } from './customSort';
+export { filteredArr } from './filteredArr';
+export { sortedArr } from './sortedArr';
