+++ src/fp/index.ts
@@ -1,15 +1,4 @@
-export {
-  lt,
-  gt,
-  gte,
-  lte,
-  inRange,
-  floorCeil,
-  isInt,
-  normalize,
-  adjustHue,
-  random
-} from './number';
-export { colorObjArr, filteredArr, sortedArr, max, min } from './array';
-export { colorObj } from './object';
-export { matchChromaChannel, expressionParser } from './string';
+export * from './number';
+export * from './array';
+export * from './object';
+export * from './string';
