+++ src/fp/number/adjustHue.ts
@@ -1,0 +1,11 @@
+// @ts-nocheck
+
+const adjustHue = (value = 0) => {
+  if (value > 0) {
+    return (value += Math.ceil(-value / 360) * 360);
+  } else {
+    return value % 360;
+  }
+};
+
+export { adjustHue };
