+++ src/fp/number/comparison.ts
@@ -1,0 +1,9 @@
+// @ts-nocheck
+
+// Comparison operators
+const gt = (x: number, y: number): boolean => x > y;
+const lt = (x: number, y: number): boolean => x < y;
+const gte = (x: number, y: number): boolean => x >= y;
+const lte = (x: number, y: number): boolean => x <= y;
+
+export { lt, gt, gte, lte };
