+++ src/fp/number/floorCeil.ts
@@ -1,0 +1,29 @@
+import { isInt } from './isInt';
+
+const { ceil, floor } = Math;
+/**
+ * @function
+ * Rounds up or down a number based on the float value.
+ * @param num The number to round up or down.
+ * @returns An integer
+ */
+const floorCeil = (num: number): number => {
+  if (isInt(num)) {
+    return num;
+  } else {
+    const strArr = num.toString().split('.');
+    const float = strArr[1];
+
+    //If the decimal value is .4  and below it will be rounded down else it will be rounded up.
+    const reFloorCeil = (float: string) => /^[0-4]$/.test(float.charAt(0));
+
+    if (reFloorCeil(float)) {
+      num = floor(num);
+    } else {
+      num = ceil(num);
+    }
+  }
+  return num;
+};
+
+export { floorCeil };
