+++ src/fp/number/index.ts
@@ -1,0 +1,10 @@
+// This module has number methods
+// @ts-nocheck
+
+export { lt, gt, gte, lte } from './comparison';
+export { random } from './random';
+export { isInt } from './isInt';
+export { normalize } from './normalize';
+export { floorCeil } from './floorCeil';
+export { adjustHue } from './adjustHue';
+export { inRange } from './inRange';
