+++ src/fp/object/customConcat.ts
@@ -1,0 +1,15 @@
+// @ts-nocheck
+
+const customConcat = (hue: object) => {
+  const res: number[] = [];
+  const { keys } = Object;
+  if (typeof hue == 'object') {
+    const hueKeys = keys(hue);
+
+    //@ts-ignore
+    res.push(...hueKeys.map((key) => hue[key]));
+  }
+  return res;
+};
+
+export { customConcat };
