+++ src/fp/object/index.ts
@@ -1,0 +1,5 @@
+// This module has object methods
+
+export { colorObj } from './colorObj';
+export { customConcat } from './customConcat';
+export { customFindKey } from './customFindKey';
