+++ src/fp/string/index.ts
@@ -1,0 +1,2 @@
+export { matchChromaChannel } from './matchChromaChannel';
+export { expressionParser } from './expressionParser';
