+++ src/palettes/adaptive.ts
@@ -1,0 +1,3 @@
+// @ts-nocheck
+
+// This module will make use of contrast ratio to create adaptive palettes
