+++ src/palettes/getHue.ts
@@ -3,9 +3,9 @@
 // @ts-nocheck
 import hueTempMap from '../color-maps/samples/hueTemperature';
 import { useMode, modeLch } from 'culori/fn';
-import { inRange } from '../fp/number.ts';
-import { min, max } from '../fp/array.ts';
-import { customConcat } from '../fp/object.ts';
+import { inRange } from '../fp/number/inRange.ts';
+import { min, max } from '../fp/array/min_max.ts';
+import { customConcat } from '../fp/object/customConcat.ts';
 import { toHex } from '../converters/toHex.ts';
 import type { Color, hue } from '../paramTypes';
