+++ src/sortBy/sortByHue.ts
@@ -1,6 +1,6 @@
 // @ts-nocheck
 import { getChannel } from '../getters_and_setters/get.ts';
-import { sortedArr } from '../fp/array.ts';
+import { sortedArr } from '../fp/array/sortedArr.ts';
 
 import type { factor, Color } from '../paramTypes.ts';
