+++ src/sortBy/sortByLightness.ts
@@ -1,7 +1,7 @@
 // @ts-nocheck
 import type { factor, Color } from '../paramTypes.ts';
 import { getChannel } from '../getters_and_setters/get.ts';
-import { sortedArr } from '../fp/array.ts';
+import { sortedArr } from '../fp/array/sortedArr.ts';
 
 /**
  * @function
